centralize elimination logic. Ensure flat orders are added for eliminated miners with open positions
